Fixed 'cast from pointer to integer of different size' error

+++ fbmmap.c
@@ -44,7 +44,7 @@ int main() {
   /* Map the device to memory. */
   fbp =
       (char *)mmap(0, screensize, PROT_READ | PROT_WRITE, MAP_SHARED, fbfd, 0);
-  if ((int)fbp == -1) {
+  if ((long)fbp == -1) {
     close(fbfd);
     perror("Failed to map framebuffer device to memory");
     exit(4);
